[core] Cleanup describeValidation
