+++ packages/x-date-pickers-pro/src/DateRangePicker/DateRangePicker.test.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import TextField from '@mui/material/TextField';
 import { describeConformance } from '@mui/monorepo/test/utils';
 import { DateRangePicker } from '@mui/x-date-pickers-pro/DateRangePicker';
-import describeValidation from '@mui/x-date-pickers-pro/tests/describeValidation';
+import { describeRangeValidation } from '@mui/x-date-pickers-pro/tests/describeRangeValidation';
 import { fireEvent, screen } from '@mui/monorepo/test/utils/createRenderer';
 import { expect } from 'chai';
 import { createPickerRenderer, stubMatchMedia, wrapPickerMount } from 'test/utils/pickers-utils';
@@ -35,11 +35,11 @@ describe('<DateRangePicker />', () => {
     }),
   );
 
-  describeValidation(DateRangePicker, () => ({
+  describeRangeValidation(DateRangePicker, () => ({
     render,
     clock,
-    isLegacyPicker: true,
-    withDate: true,
+    componentFamily: 'legacy-picker',
+    views: ['day'],
   }));
 
   it('should not open mobile picker dialog when clicked on input', () => {
