+++ packages/x-date-pickers-pro/src/MobileNextDateRangePicker/MobileNextDateRangePicker.test.tsx
@@ -3,7 +3,7 @@ import { spy } from 'sinon';
 import { expect } from 'chai';
 import { describeConformance, screen, userEvent, fireEvent } from '@mui/monorepo/test/utils';
 import { Unstable_MobileNextDateRangePicker as MobileNextDateRangePicker } from '@mui/x-date-pickers-pro/MobileNextDateRangePicker';
-import describeValidation from '@mui/x-date-pickers-pro/tests/describeValidation';
+import { describeRangeValidation } from '@mui/x-date-pickers-pro/tests/describeRangeValidation';
 import {
   wrapPickerMount,
   createPickerRenderer,
@@ -33,10 +33,11 @@ describe('<MobileNextDateRangePicker />', () => {
     ],
   }));
 
-  describeValidation(MobileNextDateRangePicker, () => ({
+  describeRangeValidation(MobileNextDateRangePicker, () => ({
     render,
     clock,
-    withDate: true,
+    componentFamily: 'new-picker',
+    views: ['day'],
   }));
 
   describe('picker state', () => {
