+++ packages/x-date-pickers-pro/src/MultiInputDateRangeField/tests/MultiInputDateRangeField.validation.test.tsx
@@ -1,13 +1,14 @@
 import { Unstable_MultiInputDateRangeField as MultiInputDateRangeField } from '@mui/x-date-pickers-pro/MultiInputDateRangeField';
-import describeValidation from '@mui/x-date-pickers-pro/tests/describeValidation';
+import { describeRangeValidation } from '@mui/x-date-pickers-pro/tests/describeRangeValidation';
 import { createPickerRenderer } from 'test/utils/pickers-utils';
 
 describe('<MultiInputDateRangeField />', () => {
   const { render, clock } = createPickerRenderer({ clock: 'fake' });
 
-  describeValidation(MultiInputDateRangeField, () => ({
+  describeRangeValidation(MultiInputDateRangeField, () => ({
     render,
     clock,
-    withDate: true,
+    componentFamily: 'field',
+    views: ['year', 'month', 'day'],
   }));
 });
