+++ packages/x-date-pickers-pro/src/MultiInputDateTimeRangeField/tests/MultiInputDateTimeRangeField.validation.test.tsx
@@ -1,14 +1,14 @@
 import { Unstable_MultiInputDateTimeRangeField as MultiInputDateTimeRangeField } from '@mui/x-date-pickers-pro/MultiInputDateTimeRangeField';
-import describeValidation from '@mui/x-date-pickers-pro/tests/describeValidation';
+import { describeRangeValidation } from '@mui/x-date-pickers-pro/tests/describeRangeValidation';
 import { createPickerRenderer } from 'test/utils/pickers-utils';
 
 describe('<MultiInputDateTimeRangeField />', () => {
   const { render, clock } = createPickerRenderer({ clock: 'fake' });
 
-  describeValidation(MultiInputDateTimeRangeField, () => ({
+  describeRangeValidation(MultiInputDateTimeRangeField, () => ({
     render,
     clock,
-    withDate: true,
-    withTime: true,
+    componentFamily: 'field',
+    views: ['year', 'month', 'day', 'hours', 'minutes'],
   }));
 });
