+++ packages/x-date-pickers-pro/src/MultiInputTimeRangeField/tests/MultiInputTimeRangeField.validation.test.tsx
@@ -1,13 +1,14 @@
 import { Unstable_MultiInputTimeRangeField as MultiInputTimeRangeField } from '@mui/x-date-pickers-pro/MultiInputTimeRangeField';
-import describeValidation from '@mui/x-date-pickers-pro/tests/describeValidation';
+import { describeRangeValidation } from '@mui/x-date-pickers-pro/tests/describeRangeValidation';
 import { createPickerRenderer } from 'test/utils/pickers-utils';
 
 describe('<MultiInputTimeRangeField />', () => {
   const { render, clock } = createPickerRenderer({ clock: 'fake' });
 
-  describeValidation(MultiInputTimeRangeField, () => ({
+  describeRangeValidation(MultiInputTimeRangeField, () => ({
     render,
     clock,
-    withTime: true,
+    componentFamily: 'field',
+    views: ['hours', 'minutes'],
   }));
 });
