+++ packages/x-date-pickers-pro/src/SingleInputDateRangeField/tests/SingleInputDateRangeField.validation.test.tsx
@@ -1,14 +1,15 @@
 import { Unstable_SingleInputDateRangeField as SingleInputDateRangeField } from '@mui/x-date-pickers-pro/SingleInputDateRangeField';
-import describeValidation from '@mui/x-date-pickers-pro/tests/describeValidation';
+import { describeRangeValidation } from '@mui/x-date-pickers-pro/tests/describeRangeValidation';
 import { createPickerRenderer } from 'test/utils/pickers-utils';
 
 describe('<SingleInputDateRangeField />', () => {
   const { render, clock } = createPickerRenderer({ clock: 'fake' });
 
-  describeValidation(SingleInputDateRangeField, () => ({
+  describeRangeValidation(SingleInputDateRangeField, () => ({
     render,
     clock,
-    withDate: true,
+    componentFamily: 'field',
+    views: ['year', 'month', 'day'],
     isSingleInput: true,
   }));
 });
