+++ packages/x-date-pickers-pro/src/tests/describeRangeValidation/index.ts
@@ -1,0 +1,1 @@
+export { describeRangeValidation } from './describeRangeValidation';
