+++ packages/x-date-pickers/src/DateCalendar/tests/DateCalendar.test.tsx
@@ -10,7 +10,7 @@ import {
   wrapPickerMount,
   createPickerRenderer,
 } from 'test/utils/pickers-utils';
-import describeValidation from '@mui/x-date-pickers/tests/describeValidation';
+import { describeValidation } from '@mui/x-date-pickers/tests/describeValidation';
 
 const isJSDOM = /jsdom/.test(window.navigator.userAgent);
 
@@ -21,7 +21,7 @@ describe('<DateCalendar />', () => {
     render,
     clock,
     views: ['year', 'month', 'day'],
-    skip: ['textField'],
+    componentFamily: 'calendar',
   }));
 
   describeConformance(<DateCalendar defaultValue={adapterToUse.date()} />, () => ({
