+++ packages/x-date-pickers/src/DateTimeField/tests/DateTimeField.validation.test.tsx
@@ -1,4 +1,4 @@
-import describeValidation from '@mui/x-date-pickers/tests/describeValidation';
+import { describeValidation } from '@mui/x-date-pickers/tests/describeValidation';
 import { Unstable_DateTimeField as DateTimeField } from '@mui/x-date-pickers/DateTimeField';
 import { createPickerRenderer } from 'test/utils/pickers-utils';
 
@@ -9,6 +9,6 @@ describe('<DateTimeField /> validation', () => {
     render,
     clock,
     views: ['year', 'month', 'day', 'hours', 'minutes'],
-    skip: ['day', 'month', 'year'],
+    componentFamily: 'field',
   }));
 });
