+++ packages/x-date-pickers/src/NextDatePicker/NextDatePicker.test.tsx
@@ -3,7 +3,7 @@ import { expect } from 'chai';
 import { Unstable_NextDatePicker as NextDatePicker } from '@mui/x-date-pickers/NextDatePicker';
 import { fireEvent, screen } from '@mui/monorepo/test/utils/createRenderer';
 import { createPickerRenderer, openPicker, stubMatchMedia } from 'test/utils/pickers-utils';
-import describeValidation from '@mui/x-date-pickers/tests/describeValidation';
+import { describeValidation } from '@mui/x-date-pickers/tests/describeValidation';
 
 const isJSDOM = /jsdom/.test(window.navigator.userAgent);
 
@@ -13,8 +13,8 @@ describe('<NextDatePicker />', () => {
   describeValidation(NextDatePicker, () => ({
     render,
     clock,
-    skip: ['textField'],
     views: ['year', 'month', 'day'],
+    componentFamily: 'new-picker',
   }));
 
   describe('prop: inputRef', () => {
