+++ packages/x-date-pickers/src/StaticDatePicker/StaticDatePicker.test.tsx
@@ -4,7 +4,7 @@ import TextField from '@mui/material/TextField';
 import { fireEvent, screen } from '@mui/monorepo/test/utils';
 import { StaticDatePicker } from '@mui/x-date-pickers/StaticDatePicker';
 import { createPickerRenderer, adapterToUse, withPickerControls } from 'test/utils/pickers-utils';
-import describeValidation from '@mui/x-date-pickers/tests/describeValidation';
+import { describeValidation } from '@mui/x-date-pickers/tests/describeValidation';
 
 const WrappedStaticDatePicker = withPickerControls(StaticDatePicker)({
   renderInput: (params) => <TextField {...params} />,
@@ -17,8 +17,7 @@ describe('<StaticDatePicker />', () => {
     render,
     clock,
     views: ['year', 'month', 'day'],
-    skip: ['textField'],
-    isLegacyPicker: true,
+    componentFamily: 'legacy-static-picker',
   }));
 
   it('render proper month', () => {
