+++ packages/x-date-pickers/src/StaticTimePicker/StaticTimePicker.test.tsx
@@ -16,7 +16,7 @@ import {
   withPickerControls,
 } from 'test/utils/pickers-utils';
 import { StaticTimePicker } from '@mui/x-date-pickers/StaticTimePicker';
-import describeValidation from '@mui/x-date-pickers/tests/describeValidation';
+import { describeValidation } from '@mui/x-date-pickers/tests/describeValidation';
 
 const WrappedStaticTimePicker = withPickerControls(StaticTimePicker)({
   renderInput: (params) => <TextField {...params} />,
@@ -29,8 +29,7 @@ describe('<StaticTimePicker />', () => {
     render,
     clock,
     views: ['hours', 'minutes'],
-    skip: ['textField'],
-    isLegacyPicker: true,
+    componentFamily: 'legacy-static-picker',
   }));
 
   describeConformance(
