+++ packages/x-date-pickers/src/tests/describeValidation/index.ts
@@ -1,0 +1,5 @@
+export { describeValidation } from './describeValidation';
+export type {
+  DescribeValidationInputOptions,
+  DescribeValidationOptions,
+} from './describeValidation.types';
